telegramsLibs/requirements.txt
aiogram==2.10.1
aiohttp==3.7.2
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
idna==2.10
multidict==5.0.0
pycparser==2.20
pyTelegramBotAPI==3.7.2
python-telegram-bot==12.8
pytz==2020.4
requests==2.24.0
six==1.15.0
tornado==6.0.4
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.6.2
